server/requirements.txt
django>=4.2,<5.0
mysqlclient>=2.1,<3.0
python-dotenv>=1.0,<2.0
django-environ>=0.9,<1.0
gunicorn>=20.0,<21.0
djangorestframework==3.14.0
django-filter==23.5
django-seed==0.3.1
psycopg2-binary==2.9.1
djangorestframework-simplejwt==5.0.0
Pillow==11.0.0